+++ src/mnspline.h
@@ -1,4 +1,4 @@
-int spline(const double *px,  const double *py, const int n, double *py2);
-int splint(const double *pxa, const double *pya, const double *py2a, const int n,
-                const double* px, double *py, const int nx);
+int spline(const double *px,  const double *py, const size_t n, double *py2);
+int splint(const double *pxa, const double *pya, const double *py2a, const size_t n,
+                const double* px, double *py, const size_t nx);
